refactor: Reservation 컨트롤러 추가 및 서비스 삭제
