+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,17 +1,21 @@
-package roomescape.service;
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+import roomescape.dto.ReservationResponseDto;
+import roomescape.model.Reservation;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.springframework.stereotype.Service;
-import roomescape.dto.ReservationResponseDto;
-import roomescape.model.Reservation;
-
-@Service
-public class ReservationService {
+@Controller
+@RequestMapping("/reservations")
+public class ReservationController {
 
     private final List<Reservation> reservations = new ArrayList<>();
 
+    @GetMapping
+    @ResponseBody
     public List<ReservationResponseDto> getReservations() {
         return reservations.stream()
                 .map(ReservationResponseDto::from)
